refactor: consolidate Yahoo sport handlers into shared factory
All 4 Yahoo sport handler files (football, baseball, basketball, hockey)
were character-identical except for baseball's extra startDate/endDate
fields in get_league_info. Extract shared handler logic into
createYahooHandlers() factory in shared/handlers.ts, reducing ~2,350
lines to ~570 + 4 thin sport wrappers (~4 lines each).

Baseball passes extraLeagueFields callback for its date fields.
Adds 28 cross-sport characterization tests covering get_league_info,
get_standings, get_roster, get_matchups, and get_free_agents.
